Added process count per champion

+++ src/bindings/champion.rs
@@ -2,6 +2,6 @@ use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
 pub struct ChampionInfo {
-    pub process_count: usize,
-    pub last_live: u32
+    pub process_count: u32,
+    pub last_live: u32,
 }
